Move ConcurrentHashSet to corresponding namespace
