+++ Mapsui/Layers/WritableLayer.cs
@@ -1,8 +1,8 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using ConcurrentCollections;
 using Mapsui.Styles;
+using Mapsui.Utilities;
 
 namespace Mapsui.Layers;
